update app module. add node controller
