+++ server/routes/posts.js
@@ -1,140 +1,17 @@
 const express = require('express');
 const router = express.Router();
-const Post = require('../models/post');
-const multer = require('multer')
+const PostController = require('../controller/post');
 const authGuard = require('../middleware/checkAuth');
+const extractFile = require('../middleware/fileCheck');
 
-// GLOBAL
-const MIME_TYPE_MAP = {
-    'image/png': 'png',
-    'image/jpeg': 'jpg',
-    'image/jpg': 'jpg'
-}
+router.get('', PostController.getAllPosts);
 
-// multer storing image uploads
-const storage = multer.diskStorage({
-    destination: (req, file, callback) => {
-        const isValid = MIME_TYPE_MAP[file.mimetype];
-        let err = new Error('Invalid File Type');
-        if(isValid) {
-            error = null
-        }
+router.post('', authGuard, extractFile, PostController.createNewPost)
 
-        callback(null, 'server/images')
-    },
-    filename: (req, file, callback) => {
-        const name = file.originalname.toLowerCase().split(' ').join('-');
-        const ext = MIME_TYPE_MAP[file.mimetype];
-        callback(null, name + '-' + Date.now() + '.' + ext);
-    }
-});
+router.put("/:id", authGuard, extractFile, PostController.updatePost)
 
+router.get("/:id", PostController.getPostById)
 
-router.get('', (req, res, next) => {
-    const pageSize = +req.query.pageSize; // + parses to number
-    const currentPage = +req.query.page;
-    const postQuery = Post.find();
-    let fetchedPosts;
-    // selected posts
-    if (pageSize && currentPage) {
-        postQuery
-            .skip(pageSize * (currentPage - 1))
-            .limit(pageSize);
-    }
-    postQuery
-        .then(documents => {
-            fetchedPosts = documents
-            return Post.count();
-        })
-        .then(count => {
-            res.status(200).json({
-                message: 'Posts fetched successfully',
-                posts: fetchedPosts,
-                maxPosts: count
-            });
-        })
-        .catch(error => {
-            res.status(500).json({
-                message: "Unable to retreive the posts"
-            })
-        })
-});
-
-router.post('', authGuard, multer({storage: storage}).single('image'), (req, res, next) => {
-    const url = req.protocol + '://' + req.get('host');
-    const post = new Post({
-        title: req.body.title,
-        content: req.body.content,
-        imagePath: url + '/images/' + req.file.filename,
-        creator: req.userData.userId
-    });
-    post.save().then(createdPost => {
-        res.status(201).json({
-            message: 'Post added',
-            post: {
-                id: createdPost._id,
-                title: createdPost.title,
-                content: createdPost.content,
-                imagePath: createdPost.imagePath
-            }
-        });
-    })
-    .catch(error => {
-        res.status(500).json({
-            message: "Unable to create the post. Please try again"
-        })
-    })
-})
-
-
-router.put("/:id", authGuard, multer({storage: storage}).single('image'), (req, res, next) => {
-    let imagePath = req.body.imagePath;
-    if (req.file) {
-        const url = req.protocol + '://' + req.get('host');        
-        imagePath = url + '/images/' + req.file.filename
-    }
-    const post = new Post({
-        _id: req.params.id,
-        title: req.body.title,
-        content: req.body.content,
-        imagePath: imagePath,
-        creator: req.userData.userId
-    })
-    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post)
-        .then(result => {
-            if (result.nModified > 0) {
-                res.status(200).json({message: "Update Successful"})
-            } else {
-                res.status(401).json({message: "Not Authorized"})
-            }
-        }).catch(error => res.status(500).json({
-            message: "Unable to update the post. Please try again"
-        }));
-})
-
-router.get("/:id", (req, res, next) => {
-    Post.findById(req.params.id).then(post => {
-        if (post) {
-            res.status(200).json(post);
-        } else {
-            res.status(404).json({message: 'Post noy found'})
-        }
-    }).catch(error => res.status(500).json({
-        message: "Unable to retrieve any posts"
-    }));
-})
-
-router.delete("/:id", authGuard, (req, res, next) => {
-    Post.deleteOne({_id: req.params.id, creator: req.userData.userId})
-    .then(result => {
-        if (result.n > 0) {
-            res.status(200).json({message: "Post Successful Deleted"})
-        } else {
-            res.status(401).json({message: "Not Authorized"})
-        }
-    }).catch(error => res.status(500).json({
-        message: "Unable to delete the post"
-    }));
-});
+router.delete("/:id", authGuard, PostController.deletePost);
 
 module.exports = router;
