+++ src/app/auth/auth.module.ts
@@ -1,0 +1,23 @@
+import { NgModule } from '@angular/core';
+import { LoginComponent } from './login/login.component';
+import { RegisterComponent } from './register/register.component';
+import { ReactiveFormsModule } from '@angular/forms';
+import { AngularMaterialModule } from '../angular-material.module';
+import { CommonModule } from '@angular/common';
+import { RouterModule } from '@angular/router';
+import { AuthRoutingModule } from './auth-routing.module';
+
+@NgModule({
+    declarations: [
+        LoginComponent,
+        RegisterComponent
+    ],
+    imports: [
+        CommonModule,
+        ReactiveFormsModule,
+        AngularMaterialModule,
+        RouterModule,
+        AuthRoutingModule
+    ]
+})
+export class AuthModule {}
