+++ src/app/guards/auth.guard.ts
@@ -16,7 +16,7 @@ export class AuthGuard implements CanActivate {
     canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): boolean | Observable<boolean> | Promise<boolean> {
         const isAuth = this.authService.getIsAuth();
         if (!isAuth) {
-            this.router.navigate(['/']);
+            this.router.navigate(['/auth/login']);
         }
         return true;
     }
