+++ src/app/services/posts.service.ts
@@ -4,7 +4,9 @@ import { Post } from '../models/Post';
 import { HttpClient } from '@angular/common/http';
 import { map } from 'rxjs/operators';
 import { Router } from '@angular/router';
+import { environment } from '../../environments/environment';
 
+const BACKEND_URL = environment.apiUrl + '/posts';
 @Injectable({
   providedIn: 'root'
 })
@@ -18,7 +20,7 @@ export class PostsService {
     // pagination query params
     const queryParams = `?pageSize=${postsPerPage}&page=${currentPage}`;
     this.httpClient
-    .get<{message: string, posts: any, maxPosts: number}>('http://localhost:3000/api/posts' + queryParams) // you must subscribe *listener
+    .get<{message: string, posts: any, maxPosts: number}>(BACKEND_URL + queryParams) // you must subscribe *listener
       .pipe(map(postData => {
         return { post: postData.posts.map(post => {
           return {
@@ -50,7 +52,7 @@ export class PostsService {
           content: string,
           imagePath: string,
           creator: string;
-        }>('http://localhost:3000/api/posts/' + id);
+        }>(BACKEND_URL + id);
   }
 
   addPost(title: string, content: string, image: File) {
@@ -58,7 +60,7 @@ export class PostsService {
     postData.append('title', title);
     postData.append('content', content);
     postData.append('image', image, title);
-    this.httpClient.post<{message: string, post: Post}>('http://localhost:3000/api/posts', postData)
+    this.httpClient.post<{message: string, post: Post}>(BACKEND_URL, postData)
       .subscribe((resData) => {
         this.router.navigate(['/']);
       });
@@ -75,14 +77,14 @@ export class PostsService {
     } else {
       postData = {id: id, title: title, content: content, imagePath: image, creator: null};
     }
-    this.httpClient.put('http://localhost:3000/api/posts/' + id, postData)
+    this.httpClient.put(BACKEND_URL + id, postData)
       .subscribe(res => {
         this.router.navigate(['/']);
       });
   }
 
   deletePost(postId: string) {
-    return this.httpClient.delete('http://localhost:3000/api/posts/' + postId);
+    return this.httpClient.delete(BACKEND_URL + '/' + postId);
   }
 
 }
